Improve sync when CPU is too slow by scrolling by more than one column
Instead of always scroling by one pixel column per timer callback and
oping that display and audio stay in sync, this change makes it check
the playing time reported by emotion and scroll the appropriate number
of columns. If the CPU is fast enough it always scrolls by one column
as before (because the timer's firing period seems very accurate).
If it's not fast enough and the timer callback that repaints the display
takes longer than the timer period, it scroll by more than one.