First step towards running in constant space
Instead of calculating and caching the whole piece, only
calculate columns that are visible on the display.
This change allows us to launch multiple calculation threads,
each of which calculates a small region, the first step to
making it run in fixed memory space instead of unlimited.

It now forgets all calculated data that is off the left edge
of the display but doesn't free stuff off the right edge yet
if you step left or return to time 0, so it doesn't run in
constant space yet.

Eventually it should precalculate ahead a limited amount so that
grey-and-repaint doesn't appear when stepping forward.