Preparatory change to scroll the screen in the main loop
So far, we have scrolled the screen in the timer callback function,
but the timer callback shouldn't call other functions. This moves the
actual scrolling into a separate function ready to try doing it by
creating an event in timer_cb() and letting the main loop do the
actual scrolling.