Check whether speclen has changed on start-calc and on result return
If we get work off the queue that is for a different speclen than the
current, the column is probably still on the screen, so calculate for
the new speclen.

If we get results from the old speclen, we remember them but don't
display them. Should we reschedule their calculation?