Optimize besseli0() function
Calculation of the Kaiser window made 211680 calls to pow(),
taking 2.22 seconds. It was doing pow(foo, k) of integer k from 1 to 24
for each sample, so this replaces that with repeated multiplication
of 1.0 by foo, resulting in a total of 641 calls to pow() in 0.01s,
speeding spettro's startup and FFT-size-change.

I've compared the outputs before and after: the maximum difference
between any two samples is 7.77156e-16 in a range from 0 to 1.0.

Timing: spettro -f 0.1 -a -e monkey.wav	 # A ten-second window on a tiny file.
before: 0.4s total CPU time. After: 0.2s