Introduce color_t and primary_t and change get_*_overlay() to return bool
Instead of munging unsigned integers, we now use color_t for colors and
primary_t for valies of R, G, B elements.
Also, get_row_overlay() and get_column_overlay() return a bool to say
whether there is an overlay at that point and fill a color_t* if so,
rather than using a funny color (0,0,0) to mean that there's no overlay there.