Lose the green line's black shadow by ORring it into the color data
On LCD screens, the green line had a black shadow
two thirds of a pixel wide, as the R and B columns were off in that column
(the color elements in my LCD pixels must be BRG or RBG.)

Now it ORs the green value to 100% when it's painted, leaving R and B alone.
Hey presto: no shadow!