Don't paint black pixels when drawing text
The caller always clears the background to black before painting text
so we can just paint the "on" pixels. This means that when text overlaps,
the result is the same whatever order the labels are drawn in, as well as
being more readable as it doesn't black out parts of previously-drawn text.