When beat lines appear, draw them, and adjust bar line width
With left and right bar lines defined, if you enabled beat lines
it used not to paint them immediately. Now it does.
Also, when changing the number of beats, repaint the old and draw the new.
Also, when beat lines appear/disappear, refresh the bar lines too so that
they change from 1 to 3 pixels wide and vice versa.
to/from 3 pixels