Add -M flag to set logmax
If you screendumped and reran spettro with the flags in the filename,
it would often come out brighter than when it was saved, requiring a
screen redraw to get the same picture. This change recods logmax in
the filename and accepts it as the parameter to a -M argument.

This also fixes some garply introduced in spettro.html by the last merge.