Reinstate icon file spettro.png to icon.png
The icon file, spettro.png, got overwritten by accident in commit 0425d026
because starting spettro up on a file and pressing 'o' with no param
changes can dump to spettro.png.

This commit reinstates the correct icon file contents and renames it to
icon.png to avoid the same mistake in future.