Refresh screen left-to-right, not right panel then left
It used to refresh the screen in most-interesting-column-first order
(i.e. from centre to right edge, then centre to left edge), but this
implies decoding compressed audio formats in reverse order for the left pane.

This makes it refresh left-to-right. which is much faster, and
promotes decoding of compressed audio files in time order, which tends
to work better too.