Iimplement calculation lookahead of 1/10th of a screenful
Now we have a simpler scheduler, it's easier to calculate a bit ahead
of the right display edge, simply scheduling the extra columns for refresh
but not displaying them when their results come in. When the display
then scrolls, the results are already in the cache.

1/10th of a display width eliminates the nasty flickering right edge
when playing and makes plain Left-Right-Arrow scrolls immediate.

Tried precalculating a screenful ahead: made response sluggish,
stealing CPU from the screen-updater.