To avoid duplicated calculations, keep a list of jobs in flight
Previously, if a second scroll event happened before the results for
the region exposed by the previous one had returned from the calc threads,
the exposed but unrepainted columns would be sent again to the calc
threads, not being in the work-to-do list or in the result cache.

By keeping track also of which calculations are in flight, we should
be able to avoid this. In practice, it reduces the number of repeated
calculations in a test piece from, for example, 136 to 4.