Implement new audio cache
Now, instead of decoding audio on-demand and keeping a copy,
we precalculate the "interesting region", i.e. any audio we might
want: the on-screen area, lookahead and lookbehind, into a pair of
buffers, one as 16-bit shorts with the original number of channels
for the audio player and one as mono floats for the FFTs, so that
fetching audio when required is immediate and requires almost no
locking. This makes audio more immmediate and means that the FFT
threads don't lock each other out, resulting in close to 100% CPU
usage when required with multiple processors and (almost) eliminating
all brief pauses in the audio player.