Merge calc_t and result_t structures as they are almost identical
result_t had only one member more than calc_t (the result of an FFT);
otherwise the fields were the same. Also, remove field "speclen" as it
can be calculated from the field fft_freq and was only used for a
debugging check of an issue that has been fixed.