Only moan when calc can't get audio if in-range, and refactor job removal
When scrolling fast, calc threads used to complain that they couldn't
get the audio they need, because it had already scrolled elsewhere.
Now they only complain if the calc is within the visible region.

Also, refactor the remove-job-in-flight code so that records of jobs
in flight are also removed when a calc fails to return a result.