When audio cache is mispositioned, copy the overlap and zero the rest
It can happen that the audio cache is mispositione if they pan in time
and a calculation thread is performed for the old region, or if they
pan by more than half a screenful and the audio thread callback happens
before half of the new audio region has been decoded.

We used to return a failure value and let the caller work it out, but
for tha audio playing thread, this made it stop playing. Now we copy the
area that *is* covered by the audio cache and zaro the rest, resulting in
a click for the audio player (which would click anyway because they just
panned) and in a duff FFT result, which is harmless.